refactor: extract retro office collision navigation system
